change candlestick to component
