+++ src/Pages/components/futures/style/all.css
@@ -91,7 +91,7 @@
     margin: 0;
     display: flex;
     width: 100%;
-    justify-content: end;
+    justify-content: flex-end;
     gap: 10px;
 }
